Folder structure cleanup
