+++ microSALT/server/app.py
@@ -1,5 +1,5 @@
 # encoding: utf-8
-from microSALT.reports.views import app
+from microSALT.server.views import app
 
 if __name__ == '__main__':
     app.run()
