+++ microSALT/store/orm_models.py
@@ -1,4 +1,4 @@
-"""Sequencing types (blast results) table definitions
+"""Samples table definition
    Heavy WIP
    By: Isak Sylvin, @sylvinite"""
 
@@ -7,6 +7,20 @@
 from sqlalchemy import *
 from sqlalchemy.orm import relationship
 from microSALT import db
+
+class Samples(db.Model):
+
+   __tablename__ = 'samples'
+   CG_ID_sample = db.Column(db.String(15), primary_key=True, nullable=False)
+   seq_types = relationship("Seq_types", back_populates="samples")
+   CG_ID_project = db.Column(db.String(15))
+   Customer_ID_sample = db.Column(db.String(15))
+   Customer_ID_project = db.Column(db.String(15))
+   date_ordered = db.Column(db.DateTime)
+   date_qc = db.Column(db.DateTime)
+   date_analysis = db.Column(db.DateTime)
+   organism = db.Column(db.String(30))
+   ST = db.Column(db.SmallInteger, default=-1)
 
 class Seq_types(db.Model):
   __tablename__ = 'seq_types'
@@ -25,3 +39,4 @@ class Seq_types(db.Model):
   contig_end = db.Column(db.Integer)
   loci_start = db.Column(db.Integer)
   loci_end = db.Column(db.Integer)
+
